mystuff/barchart-import/pyproject.toml
[tool.poetry]
name = "barchart-import"
version = "0.1.0"
description = "Python code that retrieves unusual option activity from Barchart"
authors = ["Edmondo <edmondo.porcu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
structlog = "^20.1.0"
selenium = "^3.141.0"
google-cloud-storage = "^1.29.0"
fastapi = "^0.61.0"
uvicorn = "^0.11.8"
google-cloud-pubsub = "^1.7.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pulumi = "^2.5.0"
pulumi-gcp = "^3.11.0"
pylint = "^2.5.3"
pytest = "^6.0.1"
pytest-cov = "^2.10.1"
flake8 = "^3.8.3"
isort = "^5.4.2"
autopep8 = "^1.5.4"
pytest-cases = "^2.2.1"
pre-commit = "^2.7.1"
seed-isort-config = "^2.2.0"
rootpath = "^0.1.1"
pytest-icdiff = "^0.5"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | \backup
  | buck-out
  | build
  | dist
    venv
)/
'''

[tool.poetry.scripts]
download-only = "barchart_import.importer.import_service:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

